+++ doc/changelog.rst
@@ -33,6 +33,7 @@ Addons
 Front
 ~~~~~
 
+- Fix password recovery form bug with invalid email
 - Show order reconfirmation error if product orderability changes on order
   confirmation
 - Exclude unorderable line items from basket